feat: refactor hook
